roadbotai-discord-n8n/requirements-bot.txt
discord.py==2.4.0
aiohttp==3.10.5
python-dotenv==1.0.1
audioop-lts>=0.2.1; python_version >= "3.13"
